requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
altair
annotated-types
anyio
attrs
blinker
cachetools
certifi
charset-normalizer
click
colorama
dataclasses-json
distro
faiss-cpu
frozenlist
gitdb
GitPython
greenlet
h11
httpcore
httpx
idna
Jinja2
jiter
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-experimental
langchain-openai
langchain-text-splitters
langsmith
markdown-it-py
matplotlib
MarkupSafe
marshmallow
mdurl
multidict
mistralai
mypy-extensions
narwhals
numpy
openai
orjson
packaging
pandas
pillow
protobuf
pyarrow
pydantic
pydantic_core
pydeck
Pygments
python-dateutil
python-dotenv
pytz
PyYAML
referencing
regex
requests
rich
rpds-py
six
smmap
sniffio
SQLAlchemy
streamlit
tenacity
tiktoken
toml
tornado
tqdm
typing-inspect
typing_extensions
tzdata
urllib3
watchdog
yarl
langchain_google_genai